PSF fix and new feature
Fixed PSF formula which didn't work when ni wasn't 1,
the equations for the PSF does now match the equations
presented on the web page of PSFGenerator as well
as what can be found in Born and Wolf.

Updated the dw_tiffmax so that it can be used to extract slices
with the --slice N argument. Also added the --overwrite flag.